+++ trigger/cmds.py
@@ -11,25 +11,26 @@ used as-is to execute commands and return the results as-is.
 
 __author__ = 'Jathan McCollum, Eileen Tschetter, Mark Thomas'
 __maintainer__ = 'Jathan McCollum'
-__email__ = 'jmccollum@salesforce.com'
-__copyright__ = 'Copyright 2009-2013, AOL Inc.; 2013 Salesforce.com'
-__version__ = '2.3.1'
+__email__ = 'jathan@gmail.com'
+__copyright__ = 'Copyright 2009-2013, AOL Inc.; 2014 Salesforce.com'
+__version__ = '2.4'
 
 import collections
 import datetime
 import itertools
 import os
-import sys
 from IPy import IP
 from xml.etree.cElementTree import ElementTree, Element, SubElement
+import sys
 from twisted.python import log
+from twisted.internet import defer, task
 from trigger.netdevices import NetDevices
 from trigger.conf import settings
 from trigger import exceptions
 
 
 # Exports
-__all__ = ('Commando', 'NetACLInfo')
+__all__ = ('Commando', 'ReactorlessCommando', 'NetACLInfo')
 
 
 # Default timeout in seconds for commands to return a result
@@ -508,6 +509,88 @@ class Commando(object):
             ret.append(cmd)
 
         return ret
+
+class ReactorlessCommando(Commando):
+    """
+    A reactor-less Commando subclass.
+
+    This allows multiple instances to coexist, with the side-effect that you
+    have to manage the reactor start/stop manually.
+
+    An example of how this could be used::
+
+        from twisted.internet import defer, reactor
+
+        devices = ['dev1', 'dev2']
+
+        # Our Commando instances. This is an example  to show we have two instances
+        # co-existing under the same reactor.
+        c1 = ShowClock(devices)
+        c2 = ShowUsers(devices)
+        instances = [c1, c2]
+
+        # Call the run method for each instance to get a list of Deferred task objects.
+        deferreds = []
+        for i in instances:
+            deferreds.append(i.run())
+
+        # Here we use a DeferredList to track a list of Deferred tasks that only
+        # returns once they've all completed.
+        d = defer.DeferredList(deferreds)
+
+        # Once every task has returned a result, stop the reactor
+        d.addBoth(lambda _: reactor.stop())
+
+        # And... finally, start the reactor to kick things off.
+        reactor.run()
+
+        # Inspect your results
+        print d.result
+    """
+    def _start(self):
+        """Initializes ``all_done`` instead of starting the reactor"""
+        log.msg("._start() called")
+        self.all_done = False
+
+    def _stop(self):
+        """Sets ``all_done`` to True instead of stopping the reactor"""
+        log.msg("._stop() called")
+        self.all_done = True
+
+    def run(self):
+        """
+        We've overloaded the run method to return a Deferred task object.
+        """
+        log.msg(".run() called")
+
+        # This is the default behavior
+        super(ReactorlessCommando, self).run()
+
+        # Setup a deferred to hold the delayed result and not return it until
+        # it's done. This object will be populated with the value of the
+        # results once all commands have been executed on all devices.
+        d = defer.Deferred()
+
+        # Add monitor_result as a callback
+        from twisted.internet import reactor
+        d.addCallback(self.monitor_result, reactor)
+
+        # Tell the reactor to call the callback above when it starts
+        reactor.callWhenRunning(d.callback, reactor)
+
+        return d
+
+    def monitor_result(self, result, reactor):
+        """
+        Loop periodically or until the factory stops to check if we're
+        ``all_done`` and then return the results.
+        """
+        # Once we're done, return the results
+        if self.all_done:
+            return self.results
+
+        # Otherwise tell the reactor to call me again after 0.5 seconds.
+        return task.deferLater(reactor, 0.5, self.monitor_result, result, reactor)
 
 class NetACLInfo(Commando):
     """